# openapi Диплом.yaml
openapi: 3.0.0
info:
  title: Веб приложение по контролю личных финансов
  description: Веб приложение по контролю личных финансов
  version: 1.0.0
servers:
  - url: 'http://localhost:8081/'
    description: local server
paths:
...
  /createDebt:
    post:
      description: создание долга
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createDebtInvestDto'
      responses:
        '200':
          description: Долг успешно созадн
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/200'
  /createInvest:
    post:
      description: создание инвестиции
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createDebtInvestDto'
      responses:
        '200':
          description: Долг успешно созадн
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/200'
  /detalizationDebt/{id}:
    get:
      description: детализация долга
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: детализация долга
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/detalizationDebtRespInvestDto'
  /detalizationInvest/{id}:
    get:
      description: детализация долга
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: детализация долга
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/detalizationDebtRespInvestDto'
  /detalizationDebt/edit:
    post:
      description: детализация долга
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/detalizationDebtInvestEdit'
      responses:
        '200':
          $ref: '#/components/schemas/200'
  /detalizationDebt/delete/{id}:
    delete:
      description: детализация долга
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: удаление долга
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/200'
  /detalizationInvest/edit:
    post:
      description: детализация инвестиции
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/detalizationDebtInvestEdit'
      responses:
        '200':
          description: изменение инвестиции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/200'
  /detalizationInvest/delete/{id}:
    delete:
      description: детализация долга
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: удаление инвестиции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/200'
...
components:
  schemas:
    200:
      type: object
      properties:
          message:
            type: string
            example: ok
...
    createDebtInvestDto:
      description: dto для долга и инвестиции
      type: object
      properties:
        name:
          type: string
          example: кирилл
        sum:
          type: string
          example: 5455
        dateReturn:
          type: string
          example: 23.11.2023
        accountName:
          type: string
          example: карта сбербанка
...
    detalizationDebtRespInvestDto:
      description: запрос детализации долга/инвестиции
      type: object
      properties:
        id:
          type: string
          example: 1
        name:
          type: string
          example: кирилл
        nameAccount: 
          type: string
          example: карта сбербанка
        sum:
          type: string
          example: 5455
        dateReturn:
          type: string
          example: 23.11.2023
    detalizationDebtInvestEdit:
      description: запрос детализации долга/инвестиции
      type: object
      properties:
        id:
          type: string
          example: 2
        name:
          type: string
          example: кирилл
        sum:
          type: string
          example: 5455
        dateReturn:
          type: string
          example: 23.11.2023
        dateInit:
          type: string
          example: 12.12.2023